# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/properties": {
      "get": {
        "operationId": "PropertiesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active properties (public marketplace)",
        "tags": [
          "Properties"
        ]
      },
      "post": {
        "operationId": "PropertiesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new property",
        "tags": [
          "Properties"
        ]
      }
    },
    "/api/v1/properties/mine": {
      "get": {
        "operationId": "PropertiesController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the authenticated user's properties",
        "tags": [
          "Properties"
        ]
      }
    },
    "/api/v1/properties/{id}": {
      "get": {
        "operationId": "PropertiesController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single property by id",
        "tags": [
          "Properties"
        ]
      },
      "patch": {
        "operationId": "PropertiesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an existing property",
        "tags": [
          "Properties"
        ]
      }
    },
    "/api/v1/properties/{id}/archive": {
      "post": {
        "operationId": "PropertiesController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a property",
        "tags": [
          "Properties"
        ]
      }
    },
...
    "/api/v1/mandates": {
      "post": {
        "operationId": "MandatesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMandateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a mandate (delegation to an org)",
        "tags": [
          "Mandates"
        ]
      }
    },
    "/api/v1/mandates/pending-approvals": {
      "get": {
        "operationId": "MandatesController_pending",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List pending owner approvals",
        "tags": [
          "Mandates"
        ]
      }
    },
    "/api/v1/mandates/approvals/{id}": {
      "patch": {
        "operationId": "MandatesController_decide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideApprovalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve or reject a pending approval",
        "tags": [
          "Mandates"
        ]
      }
    },
    "/api/v1/properties/{id}/media": {
      "get": {
        "operationId": "MediaController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List media for a property",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/v1/properties/{id}/media/presign": {
      "post": {
        "operationId": "MediaController_presign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignMediaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a presigned R2 upload URL",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/v1/properties/{id}/media/confirm": {
      "post": {
        "operationId": "MediaController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmMediaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm a media upload",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/v1/payments": {
      "post": {
        "operationId": "PaymentsController_initiate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initiate a payment (cash or mobile money)",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/{id}/validate": {
      "post": {
        "operationId": "PaymentsController_validate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validate a cash payment",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/my": {
      "get": {
        "operationId": "PaymentsController_myPayments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the user's own payments",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/pending-validation": {
      "get": {
        "operationId": "PaymentsController_pendingValidation",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List cash payments awaiting agent validation",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/managed": {
      "get": {
        "operationId": "PaymentsController_managed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List payments on managed properties",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/webhooks/mobile-money": {
      "post": {
        "operationId": "PaymentsController_webhook",
        "parameters": [
          {
            "name": "x-mobile-money-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mobile money provider webhook",
        "tags": [
          "Payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Paradis Immo API",
    "description": "Hybrid real estate platform API — Congo (CG). Generated snapshot.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePropertyDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePropertyDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateMandateDto": {
        "type": "object",
        "properties": {}
      },
      "DecideApprovalDto": {
        "type": "object",
        "properties": {}
      },
      "PresignMediaDto": {
        "type": "object",
        "properties": {}
      },
      "ConfirmMediaDto": {
        "type": "object",
        "properties": {}
      },
      "InitiatePaymentDto": {
        "type": "object",
        "properties": {}
      },
      "ValidatePaymentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}